# backend/clients/ts-sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Patron API",
    "description": "An open source Patreon alternative with lower fees designed for creators who publish ongoing sequential content like books, podcasts, and comics.",
    "contact": {
      "name": "Patron Team",
      "email": "support@patron.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.patron.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "List posts with cursor-based pagination and optional series filtering",
        "description": "# Errors\nReturns error if posts database query fails, series filtering fails, or connection issues occur",
        "operationId": "list_posts",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "UUID offset for cursor-based posts pagination",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of posts to return (default: 50, max: 100)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "series_id",
            "in": "query",
            "description": "Filter posts by series ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts list retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to list posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - cannot list posts for series not owned by user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during post listing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "Create a new post",
        "description": "# Errors\nReturns error if series not found, access denied, slug/number conflict, or database error",
        "operationId": "create_post",
        "requestBody": {
          "description": "Post creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid post data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to create posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - post series does not belong to authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Post creation failed - slug or number already exists in series",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during post creation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/posts/{post_id}": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get a specific post by ID with series ownership validation",
        "description": "# Errors\nReturns error if post not found, series access denied, or database connection error",
        "operationId": "get_post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "UUID of the post to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to view posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - post belongs to series not owned by user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Post not found or access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during post retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Posts"
        ],
        "summary": "Update a post",
        "description": "# Errors\nReturns error if post not found, access denied, slug/number conflict, or database update error",
        "operationId": "update_post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "UUID of the post to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Updated post data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Post updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required to update posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - cannot update post in series not owned by user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Post not found for update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Post update failed - slug or number already exists in series",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during post update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "summary": "Delete a post (soft delete) with series ownership validation",
        "description": "# Errors\nReturns error if post not found, series access denied, or database deletion error",
        "operationId": "delete_post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "UUID of the post to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Post deleted successfully"
          },
          "401": {
            "description": "Authentication required to delete posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access denied - cannot delete post in series not owned by user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Post not found for deletion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error during post deletion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePostRequest": {
        "type": "object",
        "description": "Request payload for creating a new post with content and metadata",
        "required": [
          "seriesId",
          "title",
          "content",
          "slug",
          "postNumber"
        ],
        "properties": {
          "audioFileId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional audio file attachment for the post",
            "example": "c3d4e5f6-7890-1234-cdef-123456789012"
          },
          "content": {
            "type": "string",
            "description": "Full content of the post",
            "example": "Welcome to our first episode where we discuss the fundamentals..."
          },
          "isPremium": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Restrict this post to premium subscribers only",
            "example": false
          },
          "isPublished": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Publish this post immediately upon creation",
            "example": false
          },
          "postNumber": {
            "type": "integer",
            "format": "int32",
            "description": "Episode or post number for sequencing content within the series",
            "example": 1
          },
          "seriesId": {
            "type": "string",
            "format": "uuid",
            "description": "Series to add this post to",
            "example": "b2c3d4e5-6f78-9012-bcde-f12345678901"
          },
          "slug": {
            "type": "string",
            "description": "Unique URL path segment for the new post",
            "example": "episode-1-getting-started"
          },
          "thumbnailUrl": {
            "type": [
              "string",
              "null"
            ],
            "description": "Cover image URL for the post preview",
            "example": "https://example.com/thumbnail.jpg"
          },
          "title": {
            "type": "string",
            "description": "Post title for display",
            "example": "Episode 1: Getting Started"
          },
          "videoFileId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Optional video file attachment for the post",
            "example": "d4e5f6a7-8901-2345-def0-234567890123"
          }
        },
        "example": {
          "audioFileId": "c3d4e5f6-7890-1234-cdef-123456789012",
          "content": "Welcome to our first episode where we discuss the fundamentals...",
          "isPremium": false,
          "isPublished": false,
          "postNumber": 1,
          "seriesId": "b2c3d4e5-6f78-9012-bcde-f12345678901",
          "slug": "episode-1-getting-started",
          "thumbnailUrl": "https://example.com/thumbnail.jpg",
          "title": "Episode 1: Getting Started",
          "videoFileId": "d4e5f6a7-8901-2345-def0-234567890123"
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard JSON error response structure for API endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional error code for programmatic handling",
            "example": "AUTH_INVALID_CREDENTIALS"
          },
          "error": {
            "type": "string",
            "description": "Error message describing what went wrong",
            "example": "Invalid email or password"
          }
        },
        "example": {
          "code": "AUTH_INVALID_CREDENTIALS",
          "error": "Invalid email or password"
        }
      },
...
      "PostsListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/PostResponse"
        },
        "description": "List of posts with pagination support",
        "example": [
          {
            "audioFileId": "c3d4e5f6-7890-1234-cdef-123456789012",
            "content": "Welcome to our first episode where we discuss the fundamentals...",
            "createdAt": "2023-01-01T00:00:00Z",
            "id": "d290f1ee-6c54-4b01-90e6-d701748f0851",
            "isPremium": false,
            "isPublished": true,
            "postNumber": 1,
            "seriesId": "b2c3d4e5-6f78-9012-bcde-f12345678901",
            "slug": "episode-1-getting-started",
            "thumbnailUrl": "https://example.com/thumbnail.jpg",
            "title": "Episode 1: Getting Started",
            "updatedAt": "2023-01-01T12:00:00Z",
            "videoFileId": "d4e5f6a7-8901-2345-def0-234567890123"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API key authentication using Bearer token in Authorization header"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid",
        "description": "Session-based authentication using secure HTTP-only cookies"
      }
    }
  },
...
}